Add logout method & improve login_as for ActionDispatch::IntegrationTest
This copies the Warden-related aspects of GdsSsoHelper which provides
similar methods to the Cucumber features in order to provide more
realistic login/logout functionality.

My immediate motivation for doing this is so that I can use the backend UI
to manipulate documents and then see how those changes affect the
frontend UI all within the same test. I will need to logout in the
middle of a test in order to see the latter.