Use Sidekiq fake mode by default in test environment
Sidekiq's `fake` mode[1] is its default mode for testing. It appends
jobs to a `jobs` array for inspection/processing instead of immediately
executing the worker as with the `inline` mode. I think this is
preferable because it forces us to be explicit in any tests that are
relying on Sidekiq jobs to be processed. The tests requiring the Sidekiq
jobs to be processed were updated in the previous commits in this
branch.

I think it's possible that we might see a speed improvement in the tests
with this change as there'll be less code being executed.

Note that I've chosen to clear all Sidekiq queues in the `teardown`
block of `ActiveSupport::TestCase` in test_helper.rb. This isn't
necessary for the tests to pass but the Sidekiq wiki suggests we do it
to avoid jobs remaining in the queue between tests.

Note that the default mode for Cucumber tests is still `inline`.

Note that Sidekiq's `inline` mode was enabled in
b3001c6e038dab9804d436a1971fcfdefb58bfd2 (PR #893) but it's not obvious
why.

[1]: https://github.com/mperham/sidekiq/wiki/Testing#testing-worker-queueing-fake