Implement AssetManagerStorage#size
This is the equivalent method to
AssetManagerAndQuarantinedFileStorage#size for returning the size of
the asset. In the case of AssetManagerStorage, the underlying asset is
not available on disk so we fetch the size from the Asset Manager
API. As of [1] all assets in Asset Manager have this attribute set.

We try to handle some of the things that could go wrong including an
API response without the `size` key or where the `size` is `nil`. We
also catch exceptions (such as `GdsApi::TimedOutException` or
`GdsApi::HTTPNotFound`) and send them to the notify service.

The `size` method is used in the `AttachmentData#file_is_not_empty`
validation which may be called on an instance of AttachmentData when
the underlying asset is no longer available on the disk. In this case
we only consider the attachment to be invalid if the return value of
`size` is truthy.

[1] https://github.com/alphagov/asset-manager/pull/483