Add AttachmentRedirectDueToUnpublishingIntegrationTest
For the moment this just captures the existing behaviour, i.e. it
demonstrates the scenarios in which requests for attachments are
redirected (or not) to their parent (unpublished/withdrawn) document.

I've not aimed to make it completely exhaustive, but it's enough to give
me confidence that I understand the intended behaviour. To that end, I'm
pretty sure I've identified a bug in
AttachmentVisibility#unpublished_edition which I plan to fix in a
subsequent commit. At this point, I should be able to remove the TODO
comments I've added in this commit.