Fix redirects for attachment on consultation outcome/feedback
Previously AttachmentsController#show was responding with a 404 Not
Found for attachments belonging to an unpublished consultation
outcome/feedback rather than a 302 Redirect to the parent document which
is what happens for attachments belonging to an unpublished edition
directly and I think it is the intended behaviour.

This was due to a bug in AttachmentVisibility#unpublished_edition which
didn't handle the case when an attachment belongs to a consultation
outcome/feedback, i.e. a non-edition class which includes the Attachable
module.

I've added unit tests for AttachmentVisibility#unpublished_edition to
cover the missing scenarios and then fixed the code. For completeness
I've included a test for an attachment belonging to a policy group, even
though this behaviour was already working.

Fixing the bug allowed me to remove the TODO comments in
AttachmentRedirectDueToUnpublishingIntegrationTest which now illustrates
the correct behaviour.