Explicitly use Sidekiq inline mode in govspeak_content_test
It's slightly odd to wrap the test body in the `inline` mode and to then
wrap a subset of that in `fake` mode. I'm happy to leave it for now
given that all I've done is make the previously implicit behaviour
explicit.

This is in preparation for switching the default Sidekiq mode from
`inline` to `fake` when running our tests. I temporarily switched it to
`fake` mode to discover the tests that are relying on Sidekiq to process
jobs inline, and have updated those tests to use `inline` mode
explicitly. A possible future improvement would be to avoid inline mode
and rely on explicitly draining the queue(s) instead.

This commit is best viewed with the `--ignore-all-space` argument to
`git diff`.