Explicitly use Sidekiq inline mode in document_controller_test_helpers
This is in preparation for switching the default Sidekiq mode from
`inline` to `fake` when running our tests. I temporarily switched it to
`fake` mode to discover the tests that are relying on Sidekiq to process
jobs inline, and have updated those tests to use `inline` mode
explicitly. A possible future improvement would be to avoid inline mode
and rely on explicitly draining the queue(s) instead.

This commit is best viewed with the `--ignore-all-space` argument to
`git diff`.