Update redirect URL in Asset Manager for attachments
This actually wires up the
ServiceListeners::AttachmentRedirectUrlUpdater class to receive events
whenever an edition changes state. We're actually only interested in
changes of state which result in the redirect URL of attachments being
set or unset. However, it's not trivial to work out exactly which events
those are and so for the moment we just subscribe to *all* events.

Wiring up the updater will mean that Asset Manager assets associated
with edition attachments will have their redirect URL kept up-to-date.

I want to be able to set the redirect_url for an asset in Asset Manager
in order to support the behaviour in AttachmentsController#fail which
redirects requests for attachments associated with an unpublished
edition to the parent document which should offer an explanation as to
why the document has been unpublished.

I've enhanced AttachmentRedirectDueToUnpublishingIntegrationTest which I
added in a previous commit so that it now also sets expectations for the
updates which need to be sent to the Asset Manager API. For the moment,
I've left the assertions for the orginal behaviour, because I think it
makes things a bit clearer. We'll need to remove those assertions before
we remove the AttachmentsController when we start serving attachment
assets from Asset Manager vs Whitehall.

Note that it's possible that subscribing to all events might cause an
undue number of AssetManagerUpdateAssetWorker jobs to be queued, but I'm
not convinced that that will be the case, and so I'm happy to avoid
premature optimisation until we discover that it is actually a problem.
We may also want to combine AttachmentRedirectUrlUpdater with
AttachmentDraftStatusUpdater thereby reducing the number of jobs to be
queued.