Use feedback component
This replaces the "Is there anything wrong?" link with the new feedback
component. This component asks the user if the page is useful and
reports the results to Google Analytics.

See https://github.com/alphagov/govuk_publishing_components/pull/163
for the component.

We've had to shuffle some things around to make this work:

- The page no longer needs a bottom border since it's provided by the
component
- We've had to separate #whitehall-wrapper into
`#whitehall-wrapper-slimmer` to be used by slimmer to insert the
content and `#whitehall-wrapper` which contains a lot of styling. This
prevents the `#whitehall-wrapper` styling to interfere with the
component.

https://trello.com/c/eKtXGtBD