Fix NoMethodError for #slice being called on GdsApi::Response
The Asset Manager stub was not returning a realistic value. The return
value is in fact an instance of `GdsApi::Response` not a `Hash`. This
was causing NoMethodError exceptions in integration.

I've changed the stubs to return realistic values and fixed the code in
the worker to handle them.

Note that I haven't updated stubbed values in other tests, because a
Hash will work just as well and I don't think that wrapping it in a
`GdsApi::Response` everywhere would actually gain us much.