Introduce an extra context in attachment integration test
I'm planning on extracting common setup from a bunch of these tests to
make this test more readble and more consistent with
AttachmentRedirectDueToUnpublishingIntegrationTest. All the tests except
for the one for a policy group involve an attachment being associated
with an attachable before the beginning of the test.

This commit is best viewed with the `--ignore-all-space` argument to
`git diff`.