Schedules link checker job an hour earlier
This job has begun taking longer and longer in the last week. It starts at 2AM every morning and is supposed to end at 8:30 AM.

However, we've seen it end later and later in the last 7 days, going from 8:30 to 9:45.

This interferes with a number of scheduled publishing jobs that are due to go out at 9:30 every morning.

As a first step to solving the increased amount of 5xx we're getting in whitehall admin, we are scheduling this job to start an hour earlier.

For: https://trello.com/c/aqGcNwFT/90-2-whitehall-link-checker-pegs-mysql-at-400-cpu-every-morning