Pass attachable to AttachmentData
This attempts to avoid a potential problem with the previous
implementation, in that it was possible for the
`AssetManagerCreateWhitehallAssetWorker#perform` method to be called
before the `AttachmentData`'s `Attachment` had been saved. In this case,
the asset could be created in Asset Manager without the access
limitation it should be subject to.

Because the `Attachment`'s attachable object has already been created
(through the UI) at the point the `Attachment` and `AttachmentData` are
created, we can simply pass that down to the `AttachmentData` so that it
can use it to look up the access limitation information.

One problem with this approach is that it will only work if we set the
`attachable` object on the `AttachmentData` before it's saved, so that
it's available in the `AssetManagerStorage#store!` method.

We considered failing fast in the job if the `AttachmentData`'s
`Attachment` was missing but that doesn't feel quite right as we'd be
raising exceptions for something that wasn't really exceptional
behaviour.