Extract asset_manager method in all Asset Manager worker classes
This slightly reduces duplication, but more importantly it makes the
code a bit terser and easier to read.

I'm planning to make some changes to the update worker and this will
make them easier.

Note that there's no point in memo-izing the service, because that
already happens in the Services module.