Support asset replacement in AssetManagerUpdateAssetWorker
We want to update Asset Manager whenever an attachment is replaced in
Whitehall so that Asset Manager can redirect requests for the
corresponding asset to the replacement asset. This is a first step in
that direction, although none of it is wired up yet.

These changes require the functionality in this Asset Manager PR [1].

Unfortunately, since we're not yet storing the Asset Manager asset ID in
Whitehall, we're forced to lookup both the asset being replaced and the
asset which is replacing it. This does make me wonder whether we're
making lives harder by not storing the asset ID.

[1]: https://github.com/alphagov/asset-manager/pull/499