Introduce ServiceListeners::AttachmentReplacementIdUpdater
This is similar to the other attachment-related service listeners,
but it's intended to be used to update the replacement_id of the Asset
Manager asset corresponding to a Whitehall attachment that is being
replaced. It differs slightly in that it is instantiated with an
AttachmentData rather than an Attachment. This is because I'm planning
to hook this in to the AttachmentData#replace_with! method and so we
will already know which AttachmentData the update should be performed
on.

It doesn't look as if there's currently any way to unset the
AttachmentData#replaced_by_id via the user interface and so I'm not
handling that scenario for the moment.

I'm not particularly happy with all the duplication between the various
attachment-related service listeners, but keeping the code as similar as
possible should make it easier to remove the duplication later.