Introduce Whitehall.attachment_data_notifier
This changes AttachmentData#replace_with! to publish a "replace" event
to a new notifier for AttachmentData whenever an attachment is replaced.

I plan to subscribe to these events with a handler which will update the
corresponding asset in Asset Manager with the relevant replacement_id,
so that Asset Manager can redirect requests for the replaced attachment
asset to the new attachment asset. Note that this isn't yet wired up.

It's not ideal that I'm introducing yet another notifier, but I think we
can look at simplifying the set of notifiers as a later step.