Add integration test to demonstrate attachment replacement behaviour
When an attachment file is *replaced* with another file, the
attachment's URL path changes (both the AttachmentData ID and,
potentially, the filename parts). However, requests for the original
attachment URL are refirected to the new URL by this code in
AttachmentsController#fail [1].

Note that replacing the attachment's file creates a new AttachmentData
and a new asset in Asset Manager.

This test currently just demonstrates the existing behaviour, but I plan
to add to it to test that we are updating the Asset Manager asset
corresponding to the original attachment file with a replacement ID
pointing at the asset corresponding to the new attachment file.

[1]:
https://github.com/alphagov/whitehall/blob/e65558ad9daf6091a33fc5eedc04850789fcff32/app/controllers/attachments_controller.rb#L45-L47