Update Asset Manager when attachment is replaced
This adds an extra test to AttachmentReplacementIntegrationTest to
ensure that a replacement ID is set on the Asset Manager asset
corresponding to the original attachment file.

This will mean that when we start serving Whitehall attachment requests
from Asset Manager, requests for replaced attachment assets will be
redirected to the replacement attachment asset.

I've made the test pass by subscribing to the "replace" event on the
AttachmentData notifier and handling such events using an instance of
ServiceListeners::AttachmentReplacementIdUpdater.