Create new versions of the link checker api reports tables
The existing link_checker_api_reports table currently has ~900,000 rows
and no index on the link_reportable_id and link_reportable_type columns.
This makes joining on it expensive.  We'd like to add an index on those
columns, but doing so would take a long time and block the application
because we show link check information on edition show pages.  However,
this data is:

1. quite new - we only turned on this version of the link checking code
   in the middle of january 2018.
2. quite transient - when we re-enable the overnight link checker we'll
   quickly generate link check reports for all the editions.

So it doesn't actually feel that useful to keep it around and go through
the effort of using a tool like [lhm][lhm] or [percona][percona] to peform
this change directly on the data.  We haven't removed the old version of
the table yet, because we want to apply the indexes to it so we can test
out if having those indexes on a table of that size would actually speed
things up.  We'll add those indexes manually in the console and do
some testing, and then follow up with a new migration that removes the
old tables.  We may decide to back port the existing data if we do decide
it's useful, which is why we've set the auto increment on the new versions
of the tables to start at the max id of the old versions (with a +1000 to
give us some wiggle room for any rows that get inserted after the table
rename, but before the auto increment change.

[lhm]: https://github.com/soundcloud/lhm
[percona]: https://www.percona.com/doc/percona-toolkit/2.1/pt-online-schema-change.html