Send parent_document_url to AM on publication of docs
A `Link` header, containing the URL of an attachment's parent document,
was added to Whitehall attachments in
https://github.com/alphagov/whitehall/pull/2202, so that:

> the cover sheet for an attachment is easily discoverable by crawlers and
> other clients, we should surface the attachment edition's public URL via
> HTTP headers.

We need to send the value of the attachment's parent URL to to Asset
Manager so that we can continue setting the `Link` header once we're
serving Whitehall attachments from Asset Manager.

My implementation differs slightly from that in Whitehall. In Whitehall
the `Link` header will be set irrespective of whether a document (and
therefore its attachments) is draft or public. My interpretation of PR
 #2202 is that it's unnecessary to set it for draft documents as they're
not publicly accessible to clients such as web crawlers. As such, I've
decided to send the `parent_document_url` to Asset Manager at the time a
document is published in Whitehall.

It's not obvious that it's possible to change the URL of a published
document but if it is, and if the change in URL doesn't result in a
redirect from the old to the new, then we might need to add something to
handle that.