Introduce CsvFileFromPublicHost
When we switch to storing all attachments exclusively in Asset
Manager, the CsvPreview class will no longer be able to read CSV data
from the local NFS mount to generate the previews presented to the
user[1].

Instead we can generate the CSV previews by fetching the CSV file
served at the public host. Currently that will in effect by reading
the file from the NFS mount, but when we do change over to serving
from Asset Manager it will fetch the CSV file from there.

Some CSV file attachments are almost 200Mb[2] so it is not practical
to fetch the entire file as part of the preview request. Fortunately
the preview functionality is configured to present at most 1,000 rows
and 50 columns to the user.

The CsvFileFromPublicHost class introduced in this commit fetches the
file from the public host by making a Range request for the first
MAXIMUM_RANGE_BYTES bytes. This is currently arbitrarily set at 30,000
bytes. It saves the file to a temporary file and yields the file to
the caller.

A range request always returns 206 even if the range requested is
larger than the file size. So we use this status code to check that
the data has been successfully fetched and raise an exception if not.

In a subsequent commit we will use this class in the
CsvPreviewController to build the preview.

[1] For example: https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/675936/Q4_2017_csv.csv/preview
[2]
https://www.gov.uk//government/uploads/system/uploads/attachment_data/file/399850/mappings-2015-01-27T10_26_13_00_00.csv/preview