CsvPreviewController uses csv data fetched from public host
This commit modifies the CsvPreviewController to use data fetched from
the public host by calling the CsvFileFromPublicHost class introduced
in the previous commit.

I've decided to stub the values returned from CsvFileFromPublicHost in
the CsvPreviewControllerTest but stub at the request level in the
feature spec to increase the coverage.

I'm not particularly happy with how much indirection there is in the
argument passed to CsvFileFromPublicHost. When we've switched to
AssetManagerStorage we will be able to use the `url` method instead
which should be clearer.