Use basic auth to request CSV if on integration
Assets served from
`https://www-origin.integration.publishing.service.gov.uk` are
protected behind basic authentication. For the CSV preview
functionality to work in integration Faraday needs to pass the
authentication header along with the request. The
BASIC_AUTH_CREDENTIALS environment variable is only set in integration
so we can use the presence of this variable to conditionally add the
request header.