Handle replacement of PDF attachment with non-PDF for Asset Manager
PDF attachments have thumbnail images, but non-PDF attachments do not.

Currently when a PDF attachment is replaced with a non-PDF attachment,
requests for the original PDF thumbnail are redirected to the
replacement attachment itself. This seems like a reasonably approach.

This commit fixes the equivalent Asset Manager related behaviour to
do the same thing, avoiding the following exception in
AttachmentReplacementIdUpdater#update!:

    NoMethodError: undefined method `asset_manager_path' for nil:NilClass

/cc @chrisroos