Add index on link_reportable to link_checker_api_reports
This index will speed up all access to link checker reports that starts
from an edition.  That means the documents index, edition show, and check
organisation links worker job.  This table is very large and the join is
very slow without an index on the link reportable type and id columns.
For example, with the table at ~850,000 rows it took ~15minutes on staging
to fetch 500 editions for a single org and join that list against their
link checker api reports.  Adding the index reduces this to less than a
second.  This is one of the queries that the check organistion links
worker job performs. With the index the total time to run all the check
organisation links worker jobs went from ~15 hours on staging to ~10
minutes.

Usually adding an index to a large table directly is something we'd like
to avoid, because it will lock the table for the duration.  However we
tested this on staging and it only took 3 seconds to add, so we're going
to bite the bullet and just do it.  We'll deploy this during a period of
relative minor use in the app to avoid disruption, but it's a lot easier
to apply the index directly compared to using online-schema-change tools
like [lhm][] or [percona][] to do this for us.

[lhm]:https://github.com/soundcloud/lhm
[percona]:https://www.percona.com/doc/percona-toolkit/2.1/pt-online-schema-change.html