Set a better value for MAXIMUM_RANGE_BYTES
When building the CsvPreview from the temporary file created by
CsvFileFromPublicHost we don't want to download the entire file (as
some files are almost 200Mb). Instead we make a range request for the
first MAXIMUM_RANGE_BYTES.

Setting MAXIMUM_RANGE_BYTES isn't obvious. CsvPreview will truncate
the file to the first 1000 rows and 50 columns by default, but as the
file is arranged in column then row order, we don't know how many
bytes to request to ensure that we have around 1000 rows.

I selected around 1,500 CSV files at random from those uploaded to
GOV.UK, downloaded them, truncated them to the first 1,000 rows and
counted how many bytes each file was.

Except for some very wide, very wordy CSV files[1] 99.5% of the files
sampled contained 1,000 rows within the first 300,000 bytes.

A consequence of the way that CsvPreview works is that it has no way
of knowing for the remaining 0.5% of files whether the file has been
truncated. So the truncated advice in
`app/views/csv_preview/_truncated_message` will not be shown in these
cases.

I think we can live with this in order to move us towards switching
these files to be served from Asset Manager, but we may want to
revisit this later.

[1] An example being
https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/530705/Cyber_Security_Breaches_Survey_2016_raw_data_value_labels.csv/preview. The
first 1,000 lines of this file total 8094845 bytes.