Never destroy instances of AttachmentData
There are two ways that AttachmentsController#show can respond with a
404 Not Found for a deleted attachment:

A. If the AttachmentData does not exist the call to
AttachmentData.find raises an ActiveRecord::RecordNotFound exception
which is translated into a 404 Not Found in production [1].

B. If the Attachment is marked as deleted, its edition is not included
in AttachmentVisibility#edition_ids [2] and thus
AttachmentVisibility#visible? will return `false` and this line [3] in
AttachmentsController#fail will be invoked.

Previously, for scenarios involving deleted attachments†, the application
was only relying on mechanism A when a single attachment was deleted
from a document which had never been published. Even in the very similar
scenario when a document which has never been published is discarded,
the AttachmentData instances associated with its attachments are not
destroyed.

The changes in this commmit mean that the application now relies on
mechanism B for *all* scenarios involving deleted attachments, i.e. it
makes things more consistent.

I did consider using mechanism A in all scenarios, but this would've
meant a lot more changes and it would mean we were potentially deleting
useful data.

In some ways this is a step backwards in terms of marking attachment
assets as deleted in Asset Manager, but there was already work to do on
that front and now it can start from a more consistent starting point.

† Mechanism A is still used if the AttachmentData ID in the URL path
does not exist.

[1]:
http://guides.rubyonrails.org/v5.0.6/action_controller_overview.html#rescue
[2]:
https://github.com/alphagov/whitehall/blob/32bb13cf3ae3ae9ca71fa14bd4c620c631660419/app/models/attachment_visibility.rb#L116-L118
[3]:
https://github.com/alphagov/whitehall/blob/32bb13cf3ae3ae9ca71fa14bd4c620c631660419/app/controllers/attachments_controller.rb#L31