Implement AttachmentData#accessible_to?
AttachmentVisibility#visible? conflates a number of concepts, e.g.
whether the attachment is deleted, draft, unpublished or replaced. Also
the tests for AttachmentVisibility#visible? are patchy and not very
logically organised.

This commit is a small step in test-driving a new implementation of the
logic in AttachmentVisibility#visible? via lower level predicate
methods.

My plan is to add the new methods to AttachmentData itself and to drive
out the implementation using tests in AttachmentDataVisibilityTest.

The current implementation of AttachmentVisibility seems to be overly
focussed on performance versus readability. I'm aiming to do the
reverse with the new methods, e.g. using model associations and
polymorphism in preference to more sophisticated SQL queries.