Deterministic ordering of AttachmentData#attachments
Previously we were relying on the implicit database ordering. In
practice, ordering attachments by their attachable_id shouldn't change
anything, but it makes the order deterministic and intention-revealing.

Note that the only places using AttachmentData#attachments are the
AttachmentData predicate methods that I've recently added and
AttachmentVisibility#visible_attachment which was also inherently
relying on the same ordering.