Implement AttachmentData#visible_to?
This is a replacement for AttachmentVisibility#visible?, but making use
of the predicate methods recently added to AttachmentData.

I decided to change the method signature to accept a user, because
the AttachmentData model does not have access to
GDS::SSO::ControllerMethods#current_user.

Note that I've chosen to do a lot of stubbing in the unit test for this
method, partly to avoid combinatorial explosion and partly because I
hope this method might eventually go away.