Set queue for AssetManagerAttachmentReplacementIdUpdateWorker
This service listener takes a queue argument which we previously
passed to the AssetManagerAttachmentReplacementIdUpdateWorker to set
the queue used for the AssetManagerUpdateAssetWorker. I think it makes
more sense to inline the work that AssetManagerUpdateAssetWorker does
into AssetManagerAttachmentReplacementIdUpdateWorker and therefore
execute a single job once this service listener is called. It makes
more sense to set the queue on the
AssetManagerAttachmentReplacementIdUpdateWorker.

I'll make the work asynchronous and do the inlining mentioned above in
subsequent commits.