Exclude deleted editions from AttachmentVisibility#unpublished_edition
This was originally implemented in this commit [1]. However, there's no
indication in the commit note or the unit test why deleted editions
should be included.

Unpublished documents result in a 404 Not Found, not a redirect, because
DocumentsController#find_unpublishing [2] returns `nil` if the edition is
deleted.

When an edition is deleted, its associated attachments are marked as
deleted in EditionDeleter#fire_transition! [3].

For an AttachmentData associated with an unpublished and deleted
edition, AttachmentVisibility#visible? will return `false`, both because
AttachmentVisiblity#edition_ids [4] will not include editions associated
with the AttachmentData via a deleted Attachment, and because
AttachmentVisiblity#visible_edition_scope and
AttachmentVisibility#visible_consultation_scope [5] will not include
deleted editions because of the default scope on Edition.

All this means that a request for an attachment on an unpublished and
deleted edition will currently result in a redirect to the document
which will in turn result in a 404 Not Found.

By making this change requests for an AttachmentData associated with an
unpublished and deleted edition will result in a 404 Not Found which
seems more consistent with the document behaviour and sensible from a
user's point of view.

As @chrisroos pointed out:

> I think this change means that `#unpublished_edition` will now return
> `nil` instead of raising an `ActiveRecord::RecordNotFound` exception
> in the case where the `unpublishing.edition_id` can't be found. Is
> that OK?

I think this *is* OK, because now the condition in this line of
`BaseAttachmentsController#fail` [6] will not be met and execution will
fall through to this line [7] thus still resulting in a `404 Not Found`.

[1]:
https://github.com/alphagov/whitehall/commit/b7cf7389807afec28a1da494330c1c7c34d875c4
[2]:
https://github.com/alphagov/whitehall/blob/5ec483216e3baaf95e5106c1111e5170130c8224/app/controllers/documents_controller.rb#L56-L60
[3]:
https://github.com/alphagov/whitehall/blob/5ec483216e3baaf95e5106c1111e5170130c8224/app/services/edition_deleter.rb#L20
[4]:
https://github.com/alphagov/whitehall/blob/5ec483216e3baaf95e5106c1111e5170130c8224/app/models/attachment_visibility.rb#L116-L118
[5]:
https://github.com/alphagov/whitehall/blob/master/app/models/attachment_visibility.rb#L92-L106
[6]:
https://github.com/alphagov/whitehall/blob/d509099988dd800f37f9e817eabe461e743a6ccc/app/controllers/base_attachments_controller.rb#L9
[7]:
https://github.com/alphagov/whitehall/blob/d509099988dd800f37f9e817eabe461e743a6ccc/app/controllers/base_attachments_controller.rb#L19