Set asset redirect_url to the URL of the parent document
We need to specify the full URL so that the redirect works as expected
when the asset is requested via either draft-assets or assets-origin.

I've added the `Unpublishing#document_url` method (to complement the
`Unpublishing#document_path` method) but it's not entirely clear that
this is better than calling `Whitehall.url_maker.public_document_url`
directly from within `AttachmentRedirectUrlUpdater`.

I've had to update the AttachmentRedirectDueToUnpublishingIntegrationTest
to assert that Whitehall redirects to the document path, but that we
send the full URL to Asset Manager.