Add test for request for attachment on unpublished & deleted edition
This test captures the intention behind a test which was removed from
HtmlAttachmentsControllerTest in this commit [1].

I believe the behaviour was not intended to be specific to
HtmlAttachments, i.e. that it should also apply to a FileAttachment, and
so it makes sense to re-introduce it in AttachmentsControllerTest.

Note that although the request for the attachment is redirected to the
document page, since the document has been deleted, the eventual
response to the user is a 404 Not Found.

[1]:
https://github.com/alphagov/whitehall/pull/1166/files#diff-cc68babb4d5303e466bf776b324f091bR73
[2]:
https://github.com/alphagov/whitehall/pull/719/files#diff-97a31afcb566b82d5e3725a58ed0fc37R15