Call AssetManagerUpdateAssetWorker synchronously
Adding a new job to the queue to perform the update of the assets in
asset manager seems unnecessary and may cause problems if there is a
long delay between the job in this worker being queued and the one in
AssetManagerUpdateAssetWorker. Instead we can call the logic in
AssetManagerUpdateAssetWorker using `perform`.

This change means we don't need to drain the queue in the integration
test. Note also that `perform` doesn't convert symbol arguments to
strings unlike `perform_async` so we need to change the arguments used
to call the method.