Move guard for non-existent replacement to worker
Given that some time might pass between enqueuing the
AssetManagerAttachmentReplacementIdUpdateWorker and the work being
done, it is safer to check that the AttachmentData still has a
replacement in the worker's perform method rather than here. This may
result in us enqueuing some unnecessary jobs, but seeing as these jobs
are called quite rarely this seems preferable to dealing with
hard-to-debug job failures.