Split HMCTS import into two steps: doc creation and draft creation
When doing the initial HMCTS import, just create the documents and
attachments in whitehall. Split out a separate step for running the
draft updater, which sends the document to the draft content store.

When the tasks were combined, it slowed down whitehall publishing
because these drafts were being added to the same queue as other
publications. This delayed publications on staging by up to two hours.

Having a separate draft update step lets us tune the speed at which we
create the drafts, and lets us kill the job without losing or
duplicating data in whitehall.