Simplify service listener attachment updaters
Only a FileAttachment will have an associated AttachmentData and, since
we're already checking for the presence of the AttachmentData, checking
that the attachment is a FileAttachment is redundant.

My motivation for doing this is that I want to change all these
attachment updaters to operate on an AttachmentData vs an Attachment.