Use AttachmentData#visible_edition_for in AttachmentLinkHeaderUpdater
This more closely matches the behaviour in
AttachmentsController#link_rel_headers which only adds the header if the
attachment's parent document is visible. Note that this doesn't actually
change the behaviour of the app as a whole, because the updater is
currently only wired up to handle publish/force-publish events. This
means the updater is never currently called when the parent document is
not visible, e.g. when the document is draft.

This means that we can use this updater to do a bulk update of the
assets corresponding to Whitehall attachments without adding the link
header to draft assets. While the latter probably wouldn't actually
cause any problems when we switch over to use Asset Manager to serve
Whitehall attachment assets, it would be a change in behaviour of the
system as a whole.

This change will also make it easier for me to change the updater to
operate on an AttachmentData vs an Attachment which will bring it into
line with the other updaters..