Prevent virus scanning of HMCTS forms from blocking publishing
The virus scanning process (including uploading the original file to
asset-manager) takes around 1 minute when whitehall isn't busy. But when
we upload hundreds of documents at once, as we're doing for HMCTS, a
backlog builds up and it takes over 20 minutes for each file to be
processed. This blocks publishers who are manually publishing documents.

To prevent this, periodically wait for HMCTS attachments to be uploaded.
The batch size of 10 hasn't been tuned very thoroughly, but from tesitng
it on staging it seems to be a reasonable compromise between speed of
importing the HMCTS forms and speed of uploading other attachments.