Publish destroy event to attachment notifier when attachment destroyed
I want to delete the Asset Manager assets corresponding to Whitehall
attachments when they are deleted. This is a first step towards making
this possible.

I chose to use the attachment notifier rather than the attachment data
notifier, because the deleted state that is definitely changing is on
the Attachment - the deleted state of the AttachmentData may or may not
be changing, depending on the state of other Attachments which reference
it.