Call AssetManagerUpdateAssetWorker synchronously
Now we are calling AssetManagerAttachmentReplacementIdUpdateWorker
asynchronously that worker is looking up the data it needs at job
execution time. We should therefore call AssetManagerUpdateAssetWorker
synchronously to avoid any issues with the data passed to this job
becoming stale.

Currently there are other service listeners that rely on
AssetManagerUpdateAssetWorker so it makes sense to leave it in
place. In the future we may be able to inline the relevant logic into
this worker.