Introduce AssetManagerAttachmentDeleteWorker
I'm planning to wire this up to delete assets in Asset Manager for
Whitehall attachments which are deleted.

Given the ID of an AttachmentData which is considered as deleted, this
worker will look up the AttachmentData and mark the corresponding Asset
Manager asset as deleted, as well as the asset corresponding to the
attachment's thumbnail image if appropriate.