Fix AttachmentData#significant_attachment
@kevindew pointed out that NoMethodError exceptions like this one [1]
were occurring on staging.

I discovered that this is possible if the attachment data belongs to a
deleted edition which in turn has a previous edition. I've added this
scenario to AttachmentDataVisiblityTest and called
AttachmentData#deleted? to trigger the exception.

I've fixed the problem by using the existing
AttachmentData#last_attachable method which uses the Null Object pattern
to avoid ever returning `nil`. In hindsight, I should probably have done
this in this commit [2].

[1]: https://sentry.io/govuk/app-whitehall/issues/491623681/
[2]:
https://github.com/alphagov/whitehall/commit/2256a57c603d61a0e71d69e4e7b167dd0ad1040f